+++ packages/stryker-vue-mutator/test/helpers/initSinon.ts
@@ -1,0 +1,12 @@
+import * as chai from 'chai';
+import * as chaiAsPromised from 'chai-as-promised';
+import * as sinon from 'sinon';
+import * as sinonChai from 'sinon-chai';
+chai.use(chaiAsPromised);
+chai.use(sinonChai);
+beforeEach(() => {
+  global.sandbox = sinon.createSandbox();
+});
+afterEach(() => {
+  global.sandbox.restore();
+});