+++ packages/stryker/src/utils/objectUtils.ts
@@ -93,11 +93,30 @@ export function kill(pid: number): Promise<void> {
         res();
       }
     });
-    
+
     function canIgnore(code: number | undefined) {
       // https://docs.microsoft.com/en-us/windows/desktop/Debug/system-error-codes--0-499-
       // these error codes mean the program is _already_ closed.
       return code === 255 || code === 128;
     }
   });
+}
+
+export const TimeoutExpired: unique symbol = Symbol('TimeoutExpired');
+export function timeout<T>(promise: Promise<T>, ms: number): Promise<T | typeof TimeoutExpired> {
+  const sleep = new Promise<T | typeof TimeoutExpired>((res, rej) => {
+    const timer = setTimeout(() => res(TimeoutExpired), ms);
+    promise.then(result => {
+      clearTimeout(timer);
+      res(result);
+    }).catch(error => {
+      clearTimeout(timer);
+      rej(error);
+    });
+  });
+  return sleep;
+}
+
+export function padLeft(input: string): string {
+  return input.split('\n').map(str => '\t' + str).join('\n');
 }